Added a button to change recent activity time frame, fixed last item getting cut off in half
Important: you need to run the latest version with this commit for a number of days you want to change the time frame to. For example, if you have just installed the version with this commit, you will only see the latest 1 day first, and after a week you will be able to see the whole week.