Corrected behaviour for checking the integrity of the standard lists.
SABS should now check on launch that the standard lists exist in the DB. If they do not, they will be added back and selected.

When clicking update all providers, it will only currently check whether the providers were previously selected. Not currently possible to check whether they are in the DB here as they have already been removed (ready to be re-added) by this point. This can be corrected, but I am out of time.

Tl;dr:

checkAdhellStandardPackage() will launch with MainActivity and add back any standard lists that shouldn’t be removable.

checkAdhellStandardPackage(List<String> sabsStandardSelectedPackages) is run with update all providers and will only check if the lists were selected before they were removed / re-added, not whether they existed in the DB before.

Need time to finalise this but for now it’ll do.